refactor: Refactored jinja2 filters to Nautobot standard location.
This change moves the jinja2 filters to the standard `jinja_filters.py` package that are automatically loaded by Nautobot. Also added some additional documentation for the `to_yaml` and `to_json` filters.

Fixes #6
