+++ mkdocs.yml
@@ -119,7 +119,11 @@ nav:
           - "admin/release_notes/index.md"
           - v1.0: "admin/release_notes/version_1.0.md"
           - v1.1: "admin/release_notes/version_1.1.md"
+          - v1.2: "admin/release_notes/version_1.2.md"
+          - v1.3: "admin/release_notes/version_1.3.md"
+          - v1.4: "admin/release_notes/version_1.4.md"
           - v2.0: "admin/release_notes/version_2.0.md"
+          - v2.1: "admin/release_notes/version_2.1.md"
   - Developer Guide:
       - Extending the App: "dev/extending.md"
       - Contributing to the App: "dev/contributing.md"
@@ -132,6 +136,7 @@ nav:
           - Field Descriptors: "dev/code_reference/fields.md"
           - Design Job: "dev/code_reference/design_job.md"
           - Jinja Rendering: "dev/code_reference/jinja2.md"
+          - Jinja Filters: "dev/code_reference/jinja_filters.md"
           - Template Extensions: "dev/code_reference/ext.md"
           - Util: "dev/code_reference/util.md"
   - Nautobot Docs Home ↗︎: "https://docs.nautobot.com"
